refactor: rename the utils as utility, update the README
