+++ prim3d/__init__.py
@@ -2,7 +2,7 @@
 import prim3d.libPrim3D as _C
 
 from .misc import Timer
-from .utils import (create_raycaster, marching_cubes, save_mesh)
+from .utility import (create_raycaster, marching_cubes, save_mesh)
 from .version import __version__
 
 ENABLE_OPTIX = _C.enable_optix
